# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CAFA Program Management Information System (PMIS) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/photo:
    get:
      operationId: getProfilePhoto
      tags: [profile]
      summary: Stream the signed-in user's current profile photo through an authenticated proxy
      responses:
        "200":
          description: Profile image
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
        "404":
          description: No profile photo
    post:
      operationId: completeProfilePhotoUpload
      tags: [profile]
      summary: Verify and attach an uploaded photo to the signed-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfilePhotoCompleteInput" }
      responses:
        "200":
          description: Photo attached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfilePhotoResponse" }
        "400":
          description: Invalid or expired upload descriptor
        "413":
          description: Actual uploaded size exceeds the limit
        "415":
          description: Uploaded content is not a permitted image
    delete:
      operationId: removeProfilePhoto
      tags: [profile]
      summary: Remove the signed-in user's profile photo
      responses:
        "200":
          description: Photo removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfilePhotoResponse" }
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate with username (or email) + password and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "400":
          description: Missing identifier or password
        "401":
          description: Invalid credentials
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, role, scope]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, description: "super_admin | executive_director | program_manager | senior_program_coordinator | technical_coordinator | state_office_manager | state_program_officer | viewer" }
        roleLabel: { type: string }
        scope: { type: string, description: "hq | state" }
        stateId: { type: ["integer", "null"] }
        stateName: { type: ["string", "null"] }
        stateNameAr: { type: ["string", "null"], description: "Canonical Arabic State label for locale-aware display." }
        sector: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
...
    ProfilePhotoCompleteInput:
      type: object
      required: [uploadToken]
      additionalProperties: false
      properties:
        uploadToken: { type: string }
    ProfilePhotoResponse:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl: { type: ["string", "null"], description: "Authenticated self-profile photo proxy URL; never a storage object path." }
...
    CurrentUser:
      type: object
      required: [user, permissions]
      properties:
        user: { $ref: "#/components/schemas/User" }
        permissions:
          type: array
          items: { type: string }
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string, description: "Username or email" }
        password: { type: string, format: password }
        remember: { type: boolean, default: false }
...